+++ src/features/pet-list/index.ts
@@ -1,0 +1,1 @@
+export { PetListContent } from './ui/PetListContent';
