+++ src/features/pet-list/lib/constants.ts
@@ -1,0 +1,7 @@
+export const PET_LIST_STATUS_MESSAGES: Partial<Record<number, string>> = {
+  400: '요청 값을 다시 확인해 주세요.',
+  401: '로그인이 필요한 기능입니다. 다시 로그인해 주세요.',
+  403: '반려동물 목록을 조회할 권한이 없습니다.',
+  404: '사용자 정보를 찾을 수 없습니다.',
+  500: '반려동물 목록 조회 중 서버 오류가 발생했습니다.',
+};
