+++ src/features/pet-list/ui/PetListEmptyState.tsx
@@ -1,0 +1,22 @@
+import { Link } from 'react-router-dom';
+
+export function PetListEmptyState() {
+  return (
+    <section className="flex h-full items-start">
+      <div className="w-full overflow-hidden rounded-[16px] border border-neutral-200 bg-white shadow-sm">
+        <div className="bg-[radial-gradient(circle_at_top,rgba(229,108,49,0.1),transparent_40%)] px-6 py-12 text-center sm:px-8 sm:py-14">
+          <p className="text-[16px] font-medium leading-8 text-neutral-950 sm:text-[18px]">
+            현재 등록된 반려동물이 없습니다.
+          </p>
+
+          <Link
+            to="/my/pets/new"
+            className="mt-4 inline-flex min-w-56 items-center justify-center rounded-2xl bg-brand px-8 py-4 text-[16px] font-medium text-brand-foreground transition-opacity hover:opacity-90"
+          >
+            등록하기
+          </Link>
+        </div>
+      </div>
+    </section>
+  );
+}
